fix error with message reports when custom checker uses old-style messages
when register_messages is called, it allows messages that are in the form
of 2-tuples, which means a symbol is not necessary. however, when doing
report_message_stats is called, it does not handle the case where there is
no symbol.

solution: when collecting message, use the message id as the key if the
symbol is None.

e.g.
```
Messages
--------

+------------------+------------+
|message id        |occurrences |
+==================+============+
|missing-docstring |6           |
+------------------+------------+
|W9901             |4           |
+------------------+------------+
|invalid-name      |1           |
+------------------+------------+
```