Look all the methods from ancestors for too-few-public-methods.
This patch uses two counters for too-few-public-methods and too-many-public-methods.
In order to avoid false positives, too-many-public-methods check is looking
only for methods defined in the current class. On the other hand, too-few-public-methods
should look in ancestors as well, since most of the time this check is useful only
for simple classes, without many public methods.
Closes issue #471.