Exempt str.format from redundant-keyword-arg when self is passed as a keyword
It's perfectly valid to pass self as a keyword to str.format when accessing
attributes using format's dot syntax, as in the following example:

   def __str__(self):
        return "{self.value}{self.other_value}...".format(self=self)

We're exempting only str.format with this commit, because for any other
function or method it's either impossible to do so or it's weird and
a potential sign of misdesign. Closes issue #642.