Refactor things through the imports checker
This patch transforms some public functions / methods to private
and moves some blocks of code into their own functions. Through
the latter, a couple of new messages are now emitted even though
the module couldn't be imported, such as reimported, which doesn't
make sense to not emit in this case.