Move redefined-variable-type into an extension
We got a lot of complaints about redefined-variable-type having multiple false positives
and not being reliable enough for a core checker. The fact that it doesn't take into account
duck typing makes it way to noisy when dealing with large code bases. Since it would take
some time until these concerns will be fixed, the checker becomes an extension and
can be later upgraded to a core one.