Ignore __all__ elements that don't have a parent
There used to be a bug which involved astroid and extension modules,
where the elements of the __all__ declaration didn't have a parent set.
This problem was fixed though in astroid 2.0 and we can infer properly
these declarations with the parent being set. Unfortunately, we can't
rely on astroid 2.0 right now, so the solution for not having crashes
is to skip these elements in the undefined-all-variable pattern.

Close #846