Removed the bad-builtin check
Using some of the builtins is just fine, we don't have to warn
about using filter or map, since it imposes a split opinion,
that can't be easily be agreed with in certain cases, leading to
the disabling of the check. As such, is far more suited to remove
the check.