Fix a bug where the top name of a qualified import was detected as unused variable.
The problem originates from the fact that we do not have fully qualified name
of the imports in scope locals, but only the top most one. As such, we retrieve
the full name from the node itself or the alias if there is any.
There could be a possibility of this change to emit the message about a wrong
variable when multiple imports from the same package are on the same line,
but this isn't a concern, since according to best practices, the imports
should be on different lines (so accessing the first item of the node
is by design)

Close #923.