Check proxies when verifying if an object supports a protocol
This patch adds support for understanding proxies of instances when checking
if an object, the proxy in question, supports a protocol or not. The problem
was with the isinstance call, since the proxy we have does not forward the
__instancecheck__ to its underlying wrapped object. Changing astroid for supporting
this use case requires a plethora of changes, which will lead to churn for a
small benefit.

Close #928