Run pylint from epylint using -c.
This gets rid of a potential issue which can occur if epylint was calling pylint
with lint.py instead. When this was happening, the first entry of sys.path would
have been the directory of pylint which contains lint.py. At the same time, if
someone analyzes, with epylint, a file that contains an import of a module,
and that module has the same name as one of the modules that are coming along
with pylint's lint.py, then instead of importing the module from site-packages,
the pylint's module would have been imported instead.

Close #920