Remove the support for html reporting
The html reporter was lately a second class citizen in Pylint,
being underdeveloped and mostly neglected. It has a couple of bugs,
but the more important reason why this is removed is that we now have
a much better json reporter, which can be used as a basis for more
prettier html Pylint outputs.

Part of #975