Support having plugins with options, sharing the same name
We do support plugins that have the same name, but when both of them are adding options that can be used, pylint crashes, since it already processed the options of the first plugin, introducing the plugin section in the configuration. The patch verifies that an option group already exists and if so, it does not add the same section twice, thus avoiding the duplicate section error.