Don't warn about used-before-assignment if the variable is in an exclusive except branch with NameError as a handler
Until now, we weren't emitting a warning in the case of a variable being
referenced before actual definition, with the variable access being placed
inside a TryExcept, with an except handler that caught NameError, Exception and BaseException.
While for NameError it makes sense, since it is explicit that the user is expecting
a variable to not be defined, it does not make that much sense for the rest of
the exceptions, even if they are supertypes of NameError anyway. This is because
it is not explicit and we cannot infer what the user intended to do anyway.
We are now only handling NameError.

Close #1080