Consume module level definitions by string annotations
Use of type hinting with string type names caused unused import warnings (and was very annoying)
Now, such type hints are considered as consuming names from the scope, making pylint happy.

Well... BUT...
"type:" comments and types in docstings are not considered

First one is hard to fix as comments are not available in AST(eroid).
Second is bit easier (accessible through .doc on node), but I don't need it...