We don't emit by default ``no-member`` if we have opaque inference objects in the inference results
This is controlled through the new flag ``--ignore-on-opaque-inference``, which is by
default True. The inference can return  multiple potential results while
evaluating a Python object, but some branches might not be evaluated, which
results in partial inference. In that case, it might be useful to still emit
no-member and other checks for the rest of the inferred objects.