Add a manual implementation of the edit distance algorithm
We were using editdistance, which is super fast, but the downside is that
it might require a compiler on some envs in order to be installed. As such,
it is also a hindrance not just for these kind of envs, but for PyPy as well,
since it is written in C. Let's use a handcraft implementation
for now.
Close #1423