Set min_pyver for invalid_encoding_py27 test case
This sets the version range of that test to Python versions ≥ 2.7.10 as that
is the first version for which Travis was able to run successfully.  2.7.9
failed to install the Python package, 2.7.6 and 2.7.8 failed this test.

The problem with the failed tests is the line number: older versions
apparently report line number one (1-based indexing) for an encoding error
even if the encoding specification is on the second line.