Fix re-throw `raise` statement in `_is_node_return_ended`
When processing `Raise` statement in `_is_node_return_ended` we check for
type of exception to raise. But if we are re-throwing exception which is currently handled
in `except` block there might be no exception linked to `Raise` statement.
This causes exception in `safe_infer` because we are passing `None` to it.