Refactor the checks and add the necessary boilerplate
Dropped the support for putenv as that should not be used directly, instead we just emit
about os.getenv. This check is definitely a good candidate for being replaced with typeshed
and type annotations support, but will take a bit until we'll have that, so it's good
to have until that point. Also added ChangeLog, added tests and What's new Entries.