Relaxed docstring checks for abstract methods
Abstract methods are allowed to document returns documentation even
if the default implementation does not return something.
This is so that an abstract method can document what an implementation
should return.

Abstract methods also do not need to document that they raise a
NotImplementedError. This is because a docstring author may choose to
indicate that the method is abstract in the docstring description,
which is something that we cannot detect.

Closes #2044