Includes file paths in the default report format (#2088)
The default report format in mypy, pydocstyle, and other tools includes the file path so that the user of certain terminals (e.g. iTerm) can click on each warning to open that line in their editor.

For example, here is what the default mypy output looks like:

 framework/views/billing.py:92: error: Incompatible return value type (got "View", expected 
 "DashboardTab")

With this new format, pylintoutput looks like:

  framework/views/dashboards.py:21:8: W0612: Unused variable 'x' (unused-variable)

Note that some editors (e.g. Sublime Text) index columns from 1, so this will place the cursor one column to the left of the warning.