Expose quiet mode to command line
The quiet option is parsed in Run.preprocess_options. This was necessary due to the call order in the Run class constructor. The quiet flag was also moved from OptionsManagerMixIn.init() to OptionsManagerMixin.read_config_file() because it wasn't used anywhere else in the class. Currently, quiet mode only manipulates the display of the config file path ("Using config file ").

Closes #1863