Add `--exit-zero` option (#2099)
Add a new command-line option for the use of continuous integration
scripts which abort if a command returns a non-zero status code.  If the
option is specified, and Pylint runs successfully, it will exit with 0
regardless of the number of lint issues detected.

Configuration errors, parse errors, and calling Pylint with invalid
command-line options all still return a non-zero error code, even if
`--exit-zero` is specified.

Thanks for hosting a sprint at PyCon 2018!

Close #2042