standardize help strings a bit
A number of help strings are inconsistent in trailing periods, or
spaces after them in the middle of text, or add a large amount of
whitespace by using \ to continue the string.  Standardize a lot
of these so all help messages end in a period and there is normal
whitespace in the middle.