Filter with should_analyze_file in parallel mode (#2264)
Each ChildLinter receives a file path and instantiates a linter to check
that file. As such, all files are arguments to child linters in parallel
mode. Therefore, the check for should_analyze_file must happen in the
parent linter, where knowledge of the original arguments is available.

Expand the custom should_analyze_file test to exercise parallel mode.

Close #1885