Modules with `__getattr__` are exempted by default from `no-member`
There's no easy way to figure out if a module has a particular member when
the said module uses `__getattr__`, which is a new addition to Python 3.7.
Instead we assume the safe thing to do, in the same way we do for classes,
and skip those modules from checking.

Close #2331