Simplify fixme regex
Because it is matched using search(), no need to provide leading
match-all. And the trailing part is not required either, because it
matched anything, and the regex is not anchored at end.

Also dropped the `in` optimization. Tested with a file consisting of one ~1.1MB line containing a fixme comment at the end, and saw no measurable difference in execution time
with the "in" check there or not.