Don't emit `assignment-from-no-return` for decorated function nodes
Since pylint does not currently understand decorators, it doesn't comprehend
that some functions are actually returning something when decorated (albeit there
might be a possibility for the decorator to not return anything as well).
In this case, it's safe to assume that the function might return something and
move on, which should prevent a bunch of false positives to pop up.

Close #2385