Fix internal/external distinction being broken
Partially introduced by a3d56dd, the logic for distinguishing internal
and external dependences is broken due to a failure to account for
multiple modules and packages being tested at once by pylint.

Add a mapping of module to package it belongs to in the import checker
and use this mapping to determine if an import is internal.  An import
is considered internal if the module imported is a submodule of the
package it was imported from.

Also consolidate filter logic to one function and replace the manual
caching with astroid.decorators.cached.