Correct fallback when the lines out of the parse trees are in raw_module_msgs_state
`_raw_module_msgs_state` holds a mapping of messages to lines and their status, enabled
or disabled. As such, the correct solution is to check if any of those lines matches
the current line that's out of the AST, as well as to check if it's disabled or enabled,
since it can vary depending on the pragma that was used.