Change ``unbalanced-tuple-unpacking`` back to a warning.
It used to be a warning until a couple of years ago, after it was promoted to
an error. But the check might be suggesting the wrong thing in some cases,
for instance when checking against ``sys.argv`` which cannot be known at static
analysis time. Given it might rely on potential unknown data, it's best to
have it as a warning.

Close #2522