Generalize `chained-comparison` (#2655)
`chained-comparison` detects boolean expressions such as `a < b and b < c` and suggests
it should be rewritten as `a < b < c`.

The current implementation is limited in two ways. First, cannot deal with existing chain
comparisons, so it would not detect the following:
    `a < b < c and c < d` # should be a < b < c < d

Second, the algorithm for detecting possible chains is simplistic, and would falsely
detect the following:

    `a < b < c and b < d` # no simplification possible

This change fixes both of these issues.