``assigning-non-slot`` not emitted for classes with unknown base classes.
It's possible to lint classes that don't have known bases (e.g. one of them
might come from a C extension), in which case we cannot make a lot of assumptions
about the class layout with respect to `__slots__`. As such, it's best to ignore
these classes from this check.

Close #2807