Add check for clash between arg and variable names (W1114)
It's common for a variable supplied as a positional argument to a
function or method to share the same name as the function argument. It's
rare for the variable name to match a _different_ function argument -
frequently this mismatch is due to an error: mis-ordered arguments,
missing arguments, or extra arguments supplied to the function. In the
cases where the mismatch is intentional, it is good form to supply the
argument as a keyword argument rather than a positional argument, so
that the reader does not assume that the function's argument is not
being set by the correspondingly named variable: this lint rule ignores
such cases.

Someone who is more familiar with pylint or `TypeChecker.visit_call` may
be able to significantly improve my code: I'm unfamiliar with the
important edge cases (e.g. I didn't test `functools.partial`), and half
the code is spent constructing things that might already be available
from convenience functions?