Add a command to list enabled/disabled messages
When enabling/disabling several messages and groups in a config
file, it can be unclear which messages are actually enabled and
which are disabled. This new command produces the final resolved
lists of enabled/disabled messages, sorted by symbol but with the
ID provided for use with `--help-msg`.

This commit includes a test, documentation, and adding myself to
the contributors list, as this is my first contribution.