Fixes "exception-escape" false positive with generators (#3128) (#3182)
As reported on #3128, there is a bug when the handler bound object is being used inside a generator causing a false positive. This was due to the way we were identifying if the object was "inside" its handler or not. In the previous implementation we would visit the node parents until we reached the object scope (in the generator case, the scope of the node is the generator, not the handler), while this change continues to go up the parent stack until it reaches its parent handler, or, if it's not inside its handler, it reaches the module's parent (None).

Object is considered inside its handler if it reaches it by going up in the parent chain. It's considered outside its handler if it reaches None.

Signed-off-by: Gabriel R Sezefredo <g@briel.dev>