Adds a new check 'inconsistent-quotes'.
Quoting PEP-8:

In Python, single-quoted strings and double-quoted strings are the same. This PEP does not make a recommendation for this. Pick a rule and stick to it. When a string contains single or double quote characters, however, use the other one to avoid backslashes in the string. It improves readability.

For triple-quoted strings, always use double quote characters to be consistent with the docstring convention in PEP 257.

Features:

Accounts for strings where the delimiter is swapped so an internal quote doesn't need to be escaped
Only errors on those lines that represent the module's minority delimiter.
Ignores longstrings (they could be docstrings, and checking those delimiters is someone else's responsibility)