Fixed broad_try_clause extension to check try/finally statements and to check for nested statements. (#3374)
Previously, the broad_try_except extension did not check nested statements. For example, an if/else statement and its body inside of a try clause did not result in a message. Now, the extension checks within sub-nodes (i.e., if, for, while, and with) and counts their statements too. Also, the extension also checks try/finally statements now.