Do not emit `logging-too-many-args` when `logging-format-interpolation` is emitted
The intention for `logging-too-many-args` and `logging-too-few-args` is to be emitted
when the logging string matches the expected format interpolation.
Otherwise `logging-too-many-args` will always be emitted when the logging string
does not match the expected format interpolation, especially since the arguments
have more bearing for the old style interpolation format, using modulo.

Close #3362