Protect `node_frame_class` against looping continuously on Uninferable
`node_frame_class` was receiving an `Uninferable` object. This object has the property
that its boolean value is false, but comparing it against `None` will always return true.
Because `node_frame_class` was walking up the tree to find the containing class, it was exempting
any node that was not a class, and since comparing `Uninferable` to `None` was always returning
true, the loop was running forever.

Close #3426