Continue adding the current working directory to sys.path with `fix_import_path`
This behaviour was previously removed in ac2c49867077cea9d0542560590999f2ebe00276
along with the removal of `''` to not force `pylint` load local files having the
same name as stdlib or pylint dependencies.

But we need the current working directory in sys.path in order to properly solve
relative imports.
For instance, given a package `pkg` and two modules, `pkg.A` and `pkg.B`,
pylint would not be able to import a relative import `from .B import X`
because it does not know where it can look for `.B` in the first place.
Having the current working directory in sys.path means that
`astroid.modutils.file_info_from_modpath` is still able to solve relative
imports.

At the same time, having the cwd in sys.path means that we might still load
files local to this directory during pylint's initialization.
The second part of this commit is to move `fix_import_path` before doing any
AST processing. As a side effect, this removes a nasty bug with multiple jobs
linting. Because the path was previously modified before spinning up new workers,
those new workers would have had the cwd in sys.path and as a result, they would
have been "able" to load local files from the directory, such as a malitious `astroid.py`.