Recognize explicitly abstract classes as abstract.
Related to #179 and #3098. Tweaks `utils.class_is_abstract` to not
depend purely on the presence of abstract methods, and instead also
return True for classes that either explicitly inherit from `abc.ABC`,
or explicitly define `abc.ABCMeta` as a metaclass.

This means that classes like:

    class Foo(AbstractParent, ABC):
        ...

    class Bar(AbstractParent, metaclass=ABCMeta):
        ...

no longer trigger W0223 (Method is abstract in class ... but is not
overriden).