Allow used variables to be properly consumed when different checks are enabled / disabled
When everything else was disabled, except `unused-import`, pylint was emitting `unused-import`
even though the imports in questions were used.
The emission occurred due to the fact that disabling all the messages also disabled the calling
of `visit_name`, which deals with `undefined-variable`. This resulted in `visit_name` not marking
as consumed the earlier import as expected.
This fix still allows `visit_name` to be called, but the emission of `undefined-variable` and friends
is controlled via a flag prior to emission call site.

This is somewhat of a "temporary" hack, a better solution would be to separate the emission / checking
of undefined variable from marking the variables as consumed.

Close #3445