[tox] Extend the pylint check to the relevant parts in tests
The change required would be huge, so we created a configuration
file for tests so we can enable rule progressively.
We can't add an __init__.py in test to make the command easier to launch
because there would be too much imports to modify. So we hard coded the
files and directory to lint until https://github.com/PyCQA/pylint/issues/352
is fixed.