Avoid doing a regex search for line lengths unless necessary
Before this change, we would check a line for pylint: disable-style
pragmas to determine whether to check the line length. The regex
check itself is very costly (~5% of pylint's total runtime in one
codebase), and is much more costly than the check itself.

This refactors the pylint check to instead do an approximate line
length check on everything, before using the regex to handle
exceptional, false negative cases like pragmas being the cause for the line length
overflow.

This change, in one sample codebase, lowered the check_lines
runtime from 5% of the total runtime to 0.35% of the total runtime