Experimental: Unify NamingSchemes.
The current default pylint naming schemes only differ on three parameters:

 1. characters allowed at beginning of name
 2. characters allowed elsewhere
 3. minimum length of name

Where this was not true was due to either bugs, oversight, or legacy
inertia. It's hard for me to tell the difference.  Many odd
discrepancies among the naming schemes have been rectified, and this
change will bother users by calling out some of their bizarre naming
conventions, but all such issues are valid.

For churn considerations, I've not (yet?) modified the names of the
naming scheme objects although they've transitioned from being classes
to objects, and there's a couple other bits of odd naming in here for
the same reason. My aim was chiefly to show that this was possible, and
spur consideration of the idea.

A further improvement would be to set a "minimum length" for names in
the configuration, which should be checked apart from "naming scheme"
issues. This would greatly reduce the need for users to read and
understand regular expressions, and will make messages more actionable,
but also separating that additional resonsibility would further simplify
and rectify the naming schemes.