Fix the changelog and whatsnew retroactively for 2.7 release
Remove everything that was in 2.6 changelog when it should be in 2.7
following the decision taken in #3831. This can be checked with
```
git remote add pyorigin git@github.com:PyCQA/pylint.git
git diff pyorigin/2.6 doc/whatsnew/
git diff pyorigin/2.6 ChangeLog
```
The idea is that nothing should be added between 2.6 and now in a 2.6
portion everything done between now and 2.6 should be in 2.7.