Don't reset config to defaults in parallel mode
When parallel mode is used (`--jobs 0`, `--jobs 2`, ...), the `similarities` checker seems to ignore the `min-similarity-lines` config values provided by the user. Instead it seems to use the builtin default value of 4. In "serial mode" (i.e. not specifying `--jobs`), this is not an issue.

The `reduce_map_data` method is a class method, so a new instance of the similarity checker is created. I narrowed down the issue to the fact that this instance is not configured. I'm not sure if this is the right way to fix this, but using the originally configured instance of this checker, instead of creating a new one, works.