Fix issues with Python 3.6.0 (#4446)
* tox: Enable testing python 3.6.0

The requirements_test.txt includes black, pre-commit, and pyupgrade
which are not compatible with Python 3.6.0. Add python_full_version
markers so that `tox -e py36` works with 3.6.0.

* Fix typing.Counter import for Python 3.6.0

typing.Counter was added in Python 3.6.1. In the strings checker,
guard the import with TYPE_CHECKING and use a type annotation comment
to fix usage with Python 3.6.0.

* Skip typing.NoReturn tests on Python < 3.6.2

typing.NoReturn was introduced in Python 3.6.2. Move the tests for
issue #4122 to a separate file and skip when Python is too old.

* Update ChangeLog and whatsnew