Fix matching note tags with a non-word char last
Using "\b" at the end of these patterns will only match note tags that
end in an alphanumeric character, immediately followed by a
non-alphanumeric character, or the end of the string. This is due to
"\b" being defined as a boundary between a word character ("\w") and a
non-word character ("\W"), or the end of the string. This leads to
deviations like "???" being ignored when specified.

Swapping "\b" for a positive lookahead that targets a whitespace, a
colon, or the end of a string accounts for this.

Closes #5840.