Fix pyreverse type hinting for class methods (#5881)
* Fix pyreverse type hinting for class methods

This commit fixes the alignment of arguments and their type annotations
in pyreverse printer output. It does so by checking for the type of the
current function rather than the name of the first argument.
This allows class methods having a non-standard first argument
(different from "self" or "cls") to be correctly serialized in class
diagrams.

* Add test for method with None args

According to astroid docs, this happens for builtin functions
implemented in C. In this case, we return an empty argument list.