Avoid Win32 multiprocessing limits
Creating a multiprocessing Pool with too many processes can hit
ValueError exceptions or hangs or both. The number that counts as "too
many" depends on the Python version so this change uses 56 as a
guaranteed safe limit.

Details in issue 6965

Note that this only avoids the issue if an explicit jobs count is not
passed in.