Ignore deprecated-module in access_attr_before_def_false_positive
This is because telnetlib is deprecated in Python 3.11. It's
hard to see exactly what this is testing - there's no great
explanation in-line and the test predates the first commit to
the git repo so we don't have a commit message to help.
telnetlib will be removed in 3.13, though, so at that point
we'll have to figure it out or drop the telnetlib part of the
test.

Signed-off-by: Adam Williamson <awilliam@redhat.com>