Don't emit super-init-not-called for Enum subclasses
For some reason, commit 83d544b to cpython added a `__init__`
to `Enum` which does nothing (it just says `pass`). The examples
in the Enum docs:
https://docs.python.org/3.11/howto/enum.html
still do not include calling super's `__init__` for Enum
subclasses, that define their own `__init__`, and obviously there
is no practical point to calling a method that does nothing, so
it seems best just to skip this warning for Enum.

Signed-off-by: Adam Williamson <awilliam@redhat.com>