Use trace not turtle in import_outside_toplevel test
Downstream (e.g. in Fedora) tkinter and turtle aren't always
included in the core Python package in order to avoid pulling in
a lot of graphical dependencies on systems that do not need them.
It doesn't really matter what libs are used in this test, so
let's replace it with trace.

Signed-off-by: Adam Williamson <awilliam@redhat.com>