Change syntax_error test for consistent output
Python 3.11 changes the string representation of the
SyntaxError triggered by this test - it now says "expected '('"
instead of just "invalid syntax". This changes the test to use
a different error (incomplete `for` loop) which still has just
"invalid syntax" as its description in Python 3.11. This is the
same 'bad code' used in the similar `test_stdin_syntaxerror` in
the unit tests.

Signed-off-by: Adam Williamson <awilliam@redhat.com>